+++ client/iOS/Controllers/ScreenSelectionController.h
@@ -24,8 +24,8 @@
 	NSArray *_resolution_modes;
 
 	// current selections
-	int _selection_color;
-	int _selection_resolution;
+	NSInteger _selection_color;
+	NSInteger _selection_resolution;
 }
 
 - (id)initWithConnectionParams:(ConnectionParams *)params;
